orenburg_corpus/requirements.txt
requests
lxml
python-slugify
pymystem3
